move on{Add,Remove,Toggle}Override to AreaOfStudy
